Lab 1 - Web Proxy/Game_Service/requirements.txt
grpcio
grpcio-tools
psycopg2-binary
PyJWT
requests
websockets